ml copy/requirements.txt
flask==2.2.5
numpy==1.23.5
opencv-python-headless==4.7.0.72
mediapipe==0.10.9
tensorflow==2.12.0
Pillow==9.5.0
